+++ tests/sample_taxis_test.py
@@ -1,6 +1,0 @@
-from atlas_stream import taxis
-
-
-def test_find_all_taxis():
-    results = taxis.find_all_taxis()
-    assert results.count() > 5